requirements.txt
feedparser>=6.0.0
pandas>=2.0.0
seaborn>=0.13.0
matplotlib>=3.8.0
pyyaml>=6.0
requests>=2.31.0
vnstock>=3.0.0
python-dateutil>=2.8.0
beautifulsoup4>=4.12.0
